sync-system/requirements.txt
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0

transformers>=4.36.0
sentence-transformers>=2.2.0
stable-baselines3>=2.2.0

openai>=1.0.0  
cohere>=4.37.0
anthropic>=0.8.0

gymnasium>=0.29.0
tensorboard>=2.15.0

datasets>=2.14.0
wandb>=0.16.0
ray[default]>=2.8.0

fastapi>=0.104.0
pydantic>=2.5.0
uvicorn>=0.24.0
redis>=5.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
pyjwt>=2.8.0

matplotlib>=3.7.0
seaborn>=0.13.0
scikit-learn>=1.3.0
scipy>=1.11.0
tqdm>=4.66.0
python-dotenv>=1.0.0

pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.0

httpx>=0.25.0
aiohttp>=3.9.0
tenacity>=8.2.3

gunicorn>=21.2.0
python-multipart>=0.0.6
